rearange logic and files
