+++ @xen-orchestra/web/src/components/vm/VmActionItems/VmActionPowerStateItems.vue
@@ -1,13 +1,14 @@
 <template>
-  <MenuItem icon="fa:play" :busy="isVmBusyToStart" :disabled="!isVmHalted || !canRunVmStart" @click="start">
+  <MenuItem icon="fa:play" :busy="isVmBusyToStart" :disabled="!isVmHalted" @click="start">
     {{ t('start') }}
   </MenuItem>
 </template>
 
 <script lang="ts" setup>
-import { useVmStartJob } from '@/jobs/vm-start.job'
+import { useVmStartJob } from '@/composables/vm/xo-vm-start.composable'
+import { isVmOperatingPending } from '@/utils/xo-records/vm.util'
 import MenuItem from '@core/components/menu/MenuItem.vue'
-import { VM_POWER_STATE, type XoVm } from '@vates/types'
+import { VM_OPERATIONS, VM_POWER_STATE, type XoVm } from '@vates/types'
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 
@@ -15,13 +16,16 @@ const props = defineProps<{ vm: XoVm }>()
 
 const { t } = useI18n()
 
-const isVmHalted = computed(() => props.vm.power_state === VM_POWER_STATE.HALTED)
-
 const selectedVms = computed(() => [props.vm])
 
-const { run: runVmStart, canRun: canRunVmStart, isRunning: isVmStartRunning } = useVmStartJob(selectedVms)
+const isVmHalted = computed(() => props.vm.power_state === VM_POWER_STATE.HALTED)
 
-const isVmBusyToStart = computed(() => isVmStartRunning.value)
+const { run: runVmStart, canRun: canRunVmStart } = useVmStartJob(selectedVms)
+
+const areOperationsPending = (operation: VM_OPERATIONS | VM_OPERATIONS[]) =>
+  selectedVms.value.some(vm => isVmOperatingPending(vm, operation))
+
+const isVmBusyToStart = computed(() => areOperationsPending(VM_OPERATIONS.START))
 
 const start = () => {
   if (!canRunVmStart.value || !isVmHalted.value) return
